decomp progress: 3 functions to 100%, 1 to 99%
- intelendian 2.44%->100% (asm void w/ lwbrx)
- StoreOffer ctor 59.6%->100% (page load + DateTime bitfield)
- CharCache::RecomposeCharsWithPatchIx 57%->100% (vector iter + RecomposePatches)
- DataDbgPrintLocaleSize 63.3%->99.3% (unsigned int cast for float magic)

+++ src/system/utl/Locale.cpp
@@ -21,7 +21,7 @@ DataNode DataDbgPrintLocaleSize(DataArray *) {
         "Localization table: %d entries, %gK in strings. Total mem: %gK\n",
         gDbgLocaleNumEntries,
         (float)gDbgLocaleStringsSize * 0.0009765625f,
-        (float)(gDbgLocaleStringsSize + gDbgLocaleNumEntries * 4) * 0.0009765625f
+        (float)(unsigned int)(gDbgLocaleStringsSize + gDbgLocaleNumEntries * 4) * 0.0009765625f
     );
     return 0;
 }
